Fix most warnings with gcc 4.4 on Linux
Boris Dusek <dusek@brailcom.org> writes:

> I am sending second revision of my patch, which should include
> all changes proposed and discussed so far (except for patches
> submitted by others, namely Bohdan and Chris)

> So please apply if you see no problems.

I rebased this after pushing Andrei's set of patches.
The rebase process required some minor editing, so I'm resending the
edited patch.  Would someone mind pushing it?

Also, Andrei's patch to festival introduced an implicit declaration
error.  We need to declare lisp_list_get_vect in festival_client.h.
I'll do that in another patch.

From 956082fe69f621802e28ff7fc3b862fc88db483e Mon Sep 17 00:00:00 2001
From: Boris Dusek <dusek@brailcom.org>
Date: Tue, 7 Sep 2010 14:33:05 +0200
Subject: [PATCH] Fix most warnings with gcc 4.4 on Linux.


Modified by Christopher Brannon: fix rebase conflicts.

+++ src/modules/festival.c
@@ -743,8 +743,8 @@ _festival_speak(void* nothing)
 	    }       	
 
 	    if (fwave->num_samples != 0){
-		DBG("Sending message to audio: %ld bytes\n",
-		    (fwave->num_samples) * sizeof(short));	
+		DBG("Sending message to audio: %lu bytes\n",
+		    (long unsigned) ((fwave->num_samples) * sizeof(short)));
     
 		if(FestivalDebugSaveOutput){
 		    char filename_debug[256];
@@ -951,7 +951,7 @@ cache_clean(size_t new_element_size)
     GList *gl;
     TCounterEntry *centry;
 
-    DBG("Cache: cleaning, cache size %ld kbytes (>max %d).", FestivalCache.size/1024,
+    DBG("Cache: cleaning, cache size %lu kbytes (>max %d).", (unsigned long) (FestivalCache.size/1024),
         FestivalCacheMaxKBytes);
 
     req_size = 2*FestivalCache.size/3;
